Fix 404 on the public catalog detail page
getServerSideProps fetched its own API over http://<host>. The host comes
from the request, but the scheme is hardcoded, so on an HTTPS deployment
the server sends a plaintext request to its own TLS port. That request
fails, and since any failure returns notFound, the page renders as a 404.

Reported in #495, where the same call fails with ECONNREFUSED because the
container cannot reach itself. Different trigger, same mechanism: whenever
the self-request fails for any reason, the detail page becomes a 404.

The page now reads the database directly through a shared
getPublicBookDetail() in entities/publicBook.ts, which the API route uses
as well, so the public field whitelist stays in a single place and the two
cannot drift. This matches pages/catalog/index.tsx, which already calls its
entity function in process, and it drops a needless HTTP round trip along
with the double JSON serialization.

Errors now go through errorLogger with LogEvents.API_ERROR, like the
catalog list page does.
