+++ pages/api/public/books/[id].ts
@@ -1,16 +1,9 @@
 import { PublicBookDetailType } from "@/entities/PublicBookDetailType";
-import { PublicBookType } from "@/entities/PublicBookType";
 import { prisma, reconnectPrisma } from "@/entities/db";
+import { getPublicBookDetail } from "@/entities/publicBook";
 import { LogEvents } from "@/lib/logEvents";
 import { errorLogger } from "@/lib/logger";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-const RELATED_LIMIT = 5;
-
-function parseTopics(raw: string | null | undefined): string[] {
-  if (!raw) return [];
-  return raw.split(";").map((t) => t.trim()).filter(Boolean);
-}
 
 export default async function handler(
   req: NextApiRequest,
@@ -33,84 +26,10 @@ export default async function handler(
   res.setHeader("Cache-Control", "public, max-age=60, s-maxage=300");
 
   try {
-    const book = await prisma.book.findUnique({
-      where: { id },
-      select: {
-        id: true,
-        title: true,
-        author: true,
-        isbn: true,
-        topics: true,
-        rentalStatus: true,
-        subtitle: true,
-        summary: true,
-        publisherName: true,
-        publisherDate: true,
-        pages: true,
-        minAge: true,
-        maxAge: true,
-      },
-    });
-
-    if (!book) {
+    const detail = await getPublicBookDetail(prisma, id);
+    if (!detail) {
       return res.status(404).json({ result: "Book not found" });
     }
-
-    const topics = parseTopics(book.topics);
-
-    let relatedBooks: PublicBookType[] = [];
-    if (topics.length > 0) {
-      const candidates = await prisma.book.findMany({
-        where: {
-          id: { not: id },
-          OR: topics.map((topic) => ({ topics: { contains: topic } })),
-        },
-        select: {
-          id: true,
-          title: true,
-          author: true,
-          isbn: true,
-          topics: true,
-          rentalStatus: true,
-        },
-      });
-
-      relatedBooks = candidates
-        .map((b) => ({
-          book: b,
-          shared: parseTopics(b.topics).filter((t) => topics.includes(t)).length,
-        }))
-        .sort((a, b) => b.shared - a.shared)
-        .slice(0, RELATED_LIMIT)
-        .map(({ book: b }) => ({
-          id: b.id,
-          title: b.title,
-          author: b.author,
-          isbn: b.isbn,
-          topics: b.topics,
-          rentalStatus: b.rentalStatus,
-          coverUrl: `/api/images/${b.id}`,
-        }));
-    }
-
-    const detail: PublicBookDetailType = {
-      id: book.id,
-      title: book.title,
-      author: book.author,
-      isbn: book.isbn,
-      topics: book.topics,
-      rentalStatus: book.rentalStatus,
-      coverUrl: `/api/images/${book.id}`,
-      subtitle: book.subtitle,
-      summary: book.summary,
-      publisherName: book.publisherName,
-      publisherDate: book.publisherDate,
-      pages: book.pages,
-      minAge: book.minAge,
-      maxAge: book.maxAge,
-      relatedBooks,
-    };
-
     return res.status(200).json(detail);
   } catch (error) {
     errorLogger.error(
